FastAPI/requirement.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
python-jose 
aiofiles